change to output of calcTravelTime from float to double

+++ DARP_IPS/src/utilities/MyTools.cpp
@@ -75,7 +75,7 @@ namespace Tools {
         return dist * Radius;
     }
     // function to calculate travel time
-    float calcTravelTime(double lat1, double long1, double lat2, double long2) {
+    double calcTravelTime(double lat1, double long1, double lat2, double long2) {
         double dist = calcDistance(lat1, long1, lat2, long2);
         return dist * TimePerMile;
     }
